Use size_t for length types in device/serial.
base::span will shortly require that size field is of
type size_t and is thus unsigned.

Previously the test TestHelper function created a span
with a signed length. This CL updates the test and also
the CheckReceiveError function to use size_t instead.

This ensures that when span checks are enabled in a future
CL, the code will continue to compile.

No behavior change is intended by this CL.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219872
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100709}

+++ services/device/serial/serial_io_handler_posix.h
@@ -39,10 +39,10 @@ class SerialIoHandlerPosix : public SerialIoHandler {
   bool SetControlSignals(
       const mojom::SerialHostControlSignals& control_signals) override;
   mojom::SerialConnectionInfoPtr GetPortInfo() const override;
-  int CheckReceiveError(base::span<uint8_t> buffer,
-                        int bytes_read,
-                        bool& break_detected,
-                        bool& parity_error_detected);
+  size_t CheckReceiveError(base::span<uint8_t> buffer,
+                           size_t bytes_read,
+                           bool& break_detected,
+                           bool& parity_error_detected);
 
  private:
   friend class SerialIoHandler;
@@ -70,7 +70,7 @@ class SerialIoHandlerPosix : public SerialIoHandler {
   ErrorDetectState error_detect_state_;
   bool parity_check_enabled_;
   uint8_t chars_stashed_[2];
-  int num_chars_stashed_;
+  size_t num_chars_stashed_;
 };
 
 }  // namespace device
